+++ test/integration/tests/717-sdist-test/files/package-with-failing-test/src/Lib.hs
@@ -1,0 +1,8 @@
+module Lib
+    ( someFunc
+    ) where
+
+import Language.Haskell.TH
+
+someFunc :: IO ()
+someFunc = putStrLn "aaa"