+++ test/integration/tests/717-sdist-test/files/package-with-failing-test/test/Test.hs
@@ -1,0 +1,4 @@
+main :: IO ()
+main = do
+  readFile "files/file.txt"
+  return ()