+++ test/integration/tests/717-sdist-test/files/package-with-th/src/TH.hs
@@ -1,0 +1,8 @@
+module TH (thFunc) where
+
+import Language.Haskell.TH
+
+thFunc :: Q Exp
+thFunc = runIO $ do
+  readFile "files/file.txt"
+  return $ LitE (IntegerL 5)