+++ test/integration/tests/717-sdist-test/files/subdirs/failing-in-subdir/src/LibC.hs
@@ -1,0 +1,10 @@
+{-# LANGUAGE TemplateHaskell #-}
+module LibC
+    ( someFuncC
+    ) where
+
+import THInSubdir
+import Language.Haskell.TH
+
+someFuncC :: IO ()
+someFuncC = print $(thFuncC)